Requested changes (#16)
