+++ BivittatusDB/bdb_aggregate.py
@@ -2,6 +2,9 @@ import metaclass
 try: 
     from statistics import *
     import shutil
+    import os
+    import platform
+    import time
 except:
     raise metaclass.BDBException.ImportError(f"Could not import needed files in {__file__}")
 
@@ -121,3 +124,21 @@ def LAST(table):
         return table.column[-1]
     except:
         raise metaclass.BDBException.AggregateError("Unable to perform aggregate function")
+
+#A screan cleaner with delay
+def pause_and_clean(duration):
+    time.sleep(duration)
+    
+    # Check the operating system using platform.system()
+    system_name = platform.system()
+    
+    if system_name == 'Windows':
+        os.system('cls')
+    elif system_name in ['Linux', 'Darwin']:  # Darwin' is for macOS
+        os.system('clear')
+    else:
+        raise NotImplementedError(f"The operating system '{system_name}' is not supported for clearing the screen.")
+
+# Example
+#pause_and_clean(2)  
+# Pause for 2 seconds and then wipe the screen.
